gym: hide observations while engine loading
It seems there is some asynchronous initialization that sometimes results
in observations with mostly grey background and no reward. Work-around
by attempting to detect and retry in MinetestEnv.reset().

Also in the mod have the reward initialized 1 in the mod because sometimes we were
seeing the first observation have it set to 0.