send observation for the previous action
Previously RemoteInputHandler::step() would get the action from Capnp,
fill the observation, and immediately return that observation. This
meant the observation did not reflect the action.

This commit attempts to fix that so that the observation returned
in RemoteInputHandler::step() goes into the response for the RPC
that provided the previous action.